refactor: remove the airflow folder & keeping the dags folder
