--- airflow/dags/utils/data_cleaner.py
@@ -1,9 +1,0 @@
-def clean_job_data(df):
-    """Clean and validate job data.
-    
-    Args:
-        df (pd.DataFrame): Raw job data
-        
-    Returns:
-        pd.DataFrame: Cleaned data
-    """
