--- airflow/dags/utils/logger.py
@@ -1,14 +1,0 @@
-import logging
-
-def get_logger(name):
-    """Create a logger instance with consistent formatting."""
-    logger = logging.getLogger(name)
-    if not logger.handlers:
-        handler = logging.StreamHandler()
-        formatter = logging.Formatter(
-            '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-        )
-        handler.setFormatter(formatter)
-        logger.addHandler(handler)
-        logger.setLevel(logging.INFO)
-    return logger
